checkout: allow to checkout obs imported sources of scmsync sources
This allows to skip the check of obs-scm-bridge and checks out what
is stored in OBS src server.

Note: This is not necessarly the same representation as in git
      (.obscpio instead of directories or sub modules).
      Also no modification can be committed.
      So this is only intended for people who understand the current
      internal implementation of the source storage. There is no
      guarantee that this won't change.

Checkout of entire project git is not working with this switch, as
the packagelist of osc is empty. It may also make no sense to do
this as it will not perform for large projects.